move scenarios to e2e package
