+++ e2e/opts.go
@@ -3,12 +3,11 @@ package e2e
 import (
 	"github.com/Azure/agentbaker/pkg/agent/datamodel"
 	"github.com/Azure/agentbakere2e/cluster"
-	"github.com/Azure/agentbakere2e/scenario"
 )
 
 type scenarioRunOpts struct {
 	clusterConfig *cluster.Cluster
-	scenario      *scenario.Scenario
+	scenario      *Scenario
 	nbc           *datamodel.NodeBootstrappingConfiguration
 	loggingDir    string
 }
