+++ e2e/scenario.go
@@ -1,0 +1,1134 @@
+package e2e
+
+import (
+	"fmt"
+
+	"github.com/Azure/agentbaker/pkg/agent/datamodel"
+	"github.com/Azure/agentbakere2e/cluster"
+	"github.com/Azure/agentbakere2e/config"
+	"github.com/Azure/agentbakere2e/toolkit"
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore/to"
+	"github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/compute/armcompute"
+	"github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/containerservice/armcontainerservice"
+)
+
+// This function is called internally by the scenario package to get each e2e scenario's respective config as one long slice.
+// To add a sceneario, implement a new method on the Template type in a separate file that returns a *Scenario and add
+// its return value to the slice returned by this function.
+func AllScenarios() []*Scenario {
+	return []*Scenario{
+		// block for ubuntu 1804
+		ubuntu1804(),
+		ubuntu1804gpu(),
+		ubuntu1804_azurecni(),
+		ubuntu1804gpu_azurecni(),
+		ubuntu1804ChronyRestarts(),
+
+		// block for ubuntu 2204
+		ubuntu2204(),
+		ubuntu2204ARM64(),
+		ubuntu2204CustomSysctls(),
+		ubuntu2204Wasm(),
+		ubuntu2204gpuNoDriver(),
+		ubuntu2204CustomCATrust(),
+		ubuntu2204ArtifactStreaming(),
+		ubuntu2204privatekubepkg(),
+		ubuntu2204AirGap(),
+		ubuntu2204ContainerdURL(),
+		ubuntu2204ContainerdVersion(),
+		ubuntu2204ChronyRestarts(),
+
+		// block for mariner v2
+		marinerv2(),
+		marinerv2ARM64(),
+		marinerv2gpu(),
+		marinerv2CustomSysctls(),
+		marinerv2Wasm(),
+		marinerv2_azurecni(),
+		marinerv2gpu_azurecni(),
+		marinerv2AirGap(),
+		marinerv2ARM64AirGap(),
+		marinerv2ChronyRestarts(),
+
+		// block for azurelinux v2
+		azurelinuxv2(),
+		azurelinuxv2ARM64(),
+		azurelinuxv2gpu(),
+		azurelinuxv2CustomSysctls(),
+		azurelinuxv2Wasm(),
+		azurelinuxv2_azurecni(),
+		azurelinuxv2gpu_azurecni(),
+		azurelinuxv2ARM64AirGap(),
+		azurelinuxv2AirGap(),
+		azurelinuxv2ChronyRestarts(),
+
+		// block for gpu scenarios
+		ubuntu2204gpu("ubuntu2204-gpu-ncv3", "Standard_NC6s_v3"),
+		ubuntu2204gpu("ubuntu2204-gpu-a100", "Standard_NC24ads_A100_v4"),
+		ubuntu2204gpu("ubuntu2204-gpu-a10", "Standard_NV6ads_A10_v5"),
+		ubuntu2204GPUGridDriver(),
+	}
+}
+
+func azurelinuxv2() *Scenario {
+	return &Scenario{
+		Name:        "azurelinuxv2",
+		Description: "Tests that a node using a AzureLinuxV2 (CgroupV2) VHD can be properly bootstrapped",
+		Tags: Tags{
+			Name:     "azurelinuxv2",
+			OS:       "azurelinuxv2",
+			Platform: "x64",
+		},
+		Config: Config{
+			Cluster:     cluster.ClusterKubenet,
+			VHDSelector: config.VHDAzureLinuxV2Gen2,
+			BootstrapConfigMutator: func(nbc *datamodel.NodeBootstrappingConfiguration) {
+				nbc.ContainerService.Properties.AgentPoolProfiles[0].Distro = "aks-azurelinux-v2-gen2"
+				nbc.AgentPoolProfile.Distro = "aks-azurelinux-v2-gen2"
+			},
+		},
+	}
+}
+
+func azurelinuxv2AirGap() *Scenario {
+	return &Scenario{
+		Name:        "azurelinuxv2-airgap",
+		Description: "Tests that a node using a AzureLinuxV2 (CgroupV2) VHD can be properly bootstrapped",
+		Tags: Tags{
+			Name:     "azurelinuxv2-airgap",
+			OS:       "azurelinuxv2",
+			Platform: "x64",
+			Airgap:   true,
+		},
+		Config: Config{
+			Cluster:     cluster.ClusterKubenetAirgap,
+			VHDSelector: config.VHDAzureLinuxV2Gen2,
+			BootstrapConfigMutator: func(nbc *datamodel.NodeBootstrappingConfiguration) {
+				nbc.ContainerService.Properties.AgentPoolProfiles[0].Distro = "aks-azurelinux-v2-gen2"
+				nbc.AgentPoolProfile.Distro = "aks-azurelinux-v2-gen2"
+			},
+			Airgap: true,
+		},
+	}
+}
+
+func azurelinuxv2ARM64() *Scenario {
+	return &Scenario{
+		Name:        "azurelinuxv2-arm64",
+		Description: "Tests that a node using a AzureLinuxV2 (CgroupV2) VHD on ARM64 architecture can be properly bootstrapped",
+		Tags: Tags{
+			Name:     "azurelinuxv2-arm64",
+			OS:       "azurelinuxv2",
+			Platform: "arm64",
+		},
+		Config: Config{
+			Cluster:     cluster.ClusterKubenet,
+			VHDSelector: config.VHDAzureLinuxV2Gen2Arm64,
+			BootstrapConfigMutator: func(nbc *datamodel.NodeBootstrappingConfiguration) {
+				nbc.ContainerService.Properties.AgentPoolProfiles[0].VMSize = "Standard_D2pds_V5"
+				nbc.ContainerService.Properties.AgentPoolProfiles[0].Distro = "aks-azurelinux-v2-arm64-gen2"
+				nbc.ContainerService.Properties.OrchestratorProfile.KubernetesConfig.CustomKubeBinaryURL = "https://acs-mirror.azureedge.net/kubernetes/v1.24.9/binaries/kubernetes-node-linux-arm64.tar.gz"
+				nbc.AgentPoolProfile.VMSize = "Standard_D2pds_V5"
+				nbc.AgentPoolProfile.Distro = "aks-azurelinux-v2-arm64-gen2"
+				nbc.IsARM64 = true
+			},
+			VMConfigMutator: func(vmss *armcompute.VirtualMachineScaleSet) {
+				vmss.SKU.Name = to.Ptr("Standard_D2pds_V5")
+			},
+		},
+	}
+}
+
+func azurelinuxv2ARM64AirGap() *Scenario {
+	return &Scenario{
+		Name:        "azurelinuxv2-arm64-airgap",
+		Description: "Tests that a node using a AzureLinuxV2 (CgroupV2) VHD on ARM64 architecture can be properly bootstrapped",
+		Tags: Tags{
+			Name:     "azurelinuxv2-arm64-airgap",
+			OS:       "azurelinuxv2",
+			Platform: "arm64",
+			Airgap:   true,
+		},
+		Config: Config{
+			Cluster:     cluster.ClusterKubenetAirgap,
+			VHDSelector: config.VHDAzureLinuxV2Gen2Arm64,
+			BootstrapConfigMutator: func(nbc *datamodel.NodeBootstrappingConfiguration) {
+				nbc.ContainerService.Properties.AgentPoolProfiles[0].VMSize = "Standard_D2pds_V5"
+				nbc.ContainerService.Properties.AgentPoolProfiles[0].Distro = "aks-azurelinux-v2-arm64-gen2"
+				nbc.ContainerService.Properties.OrchestratorProfile.KubernetesConfig.CustomKubeBinaryURL = "https://acs-mirror.azureedge.net/kubernetes/v1.24.9/binaries/kubernetes-node-linux-arm64.tar.gz"
+				nbc.AgentPoolProfile.VMSize = "Standard_D2pds_V5"
+				nbc.AgentPoolProfile.Distro = "aks-azurelinux-v2-arm64-gen2"
+				nbc.IsARM64 = true
+			},
+			VMConfigMutator: func(vmss *armcompute.VirtualMachineScaleSet) {
+				vmss.SKU.Name = to.Ptr("Standard_D2pds_V5")
+			},
+			Airgap: true,
+		},
+	}
+}
+
+func azurelinuxv2_azurecni() *Scenario {
+	return &Scenario{
+		Name:        "azurelinuxv2-azurecni",
+		Description: "azurelinuxv2 scenario on a cluster configured with Azure CNI",
+		Tags: Tags{
+			Name:     "azurelinuxv2-azurecni",
+			OS:       "azurelinuxv2",
+			Platform: "x64",
+		},
+		Config: Config{
+			Cluster:     cluster.ClusterAzureNetwork,
+			VHDSelector: config.VHDAzureLinuxV2Gen2,
+			BootstrapConfigMutator: func(nbc *datamodel.NodeBootstrappingConfiguration) {
+				nbc.ContainerService.Properties.OrchestratorProfile.KubernetesConfig.NetworkPlugin = string(armcontainerservice.NetworkPluginAzure)
+				nbc.AgentPoolProfile.KubernetesConfig.NetworkPlugin = string(armcontainerservice.NetworkPluginAzure)
+				nbc.ContainerService.Properties.AgentPoolProfiles[0].Distro = "aks-azurelinux-v2-gen2"
+				nbc.AgentPoolProfile.Distro = "aks-azurelinux-v2-gen2"
+			},
+		},
+	}
+}
+
+func azurelinuxv2ChronyRestarts() *Scenario {
+	return &Scenario{
+		Name:        "azurelinuxv2-chrony-restarts",
+		Description: "Tests that the chrony service restarts if it is killed",
+		Tags: Tags{
+			Name:     "azurelinuxv2-chrony-restarts",
+			OS:       "azurelinuxv2",
+			Platform: "x64",
+		},
+		Config: Config{
+			Cluster:     cluster.ClusterKubenet,
+			VHDSelector: config.VHDAzureLinuxV2Gen2,
+			BootstrapConfigMutator: func(nbc *datamodel.NodeBootstrappingConfiguration) {
+				nbc.ContainerService.Properties.AgentPoolProfiles[0].Distro = "aks-azurelinux-v2-gen2"
+				nbc.AgentPoolProfile.Distro = "aks-azurelinux-v2-gen2"
+			},
+			LiveVMValidators: []*LiveVMValidator{
+				serviceCanRestartValidator("chronyd", 10),
+				FileHasContentsValidator("/etc/systemd/system/chronyd.service.d/10-chrony-restarts.conf", "Restart=always"),
+				FileHasContentsValidator("/etc/systemd/system/chronyd.service.d/10-chrony-restarts.conf", "RestartSec=5"),
+			},
+		},
+	}
+}
+
+func azurelinuxv2CustomSysctls() *Scenario {
+	customSysctls := map[string]string{
+		"net.ipv4.ip_local_port_range":       "32768 62535",
+		"net.netfilter.nf_conntrack_max":     "2097152",
+		"net.netfilter.nf_conntrack_buckets": "524288",
+		"net.ipv4.tcp_keepalive_intvl":       "90",
+		"net.ipv4.ip_local_reserved_ports":   "",
+	}
+	customContainerdUlimits := map[string]string{
+		"LimitMEMLOCK": "75000",
+		"LimitNOFILE":  "1048",
+	}
+	return &Scenario{
+		Name:        "azurelinuxv2-custom-sysctls",
+		Description: "tests that a AzureLinuxV2 (CgroupV2) VHD can be properly bootstrapped when supplied custom node config that contains custom sysctl settings",
+		Tags: Tags{
+			Name:     "azurelinuxv2-custom-sysctls",
+			OS:       "azurelinuxv2",
+			Platform: "x64",
+		},
+		Config: Config{
+			Cluster:     cluster.ClusterKubenet,
+			VHDSelector: config.VHDAzureLinuxV2Gen2,
+			BootstrapConfigMutator: func(nbc *datamodel.NodeBootstrappingConfiguration) {
+				customLinuxConfig := &datamodel.CustomLinuxOSConfig{
+					Sysctls: &datamodel.SysctlConfig{
+						NetNetfilterNfConntrackMax:     to.Ptr(toolkit.StrToInt32(customSysctls["net.netfilter.nf_conntrack_max"])),
+						NetNetfilterNfConntrackBuckets: to.Ptr(toolkit.StrToInt32(customSysctls["net.netfilter.nf_conntrack_buckets"])),
+						NetIpv4IpLocalPortRange:        customSysctls["net.ipv4.ip_local_port_range"],
+						NetIpv4TcpkeepaliveIntvl:       to.Ptr(toolkit.StrToInt32(customSysctls["net.ipv4.tcp_keepalive_intvl"])),
+					},
+					UlimitConfig: &datamodel.UlimitConfig{
+						MaxLockedMemory: customContainerdUlimits["LimitMEMLOCK"],
+						NoFile:          customContainerdUlimits["LimitNOFILE"],
+					},
+				}
+				nbc.AgentPoolProfile.CustomLinuxOSConfig = customLinuxConfig
+				nbc.ContainerService.Properties.AgentPoolProfiles[0].CustomLinuxOSConfig = customLinuxConfig
+				nbc.ContainerService.Properties.AgentPoolProfiles[0].Distro = "aks-azurelinux-v2-gen2"
+				nbc.AgentPoolProfile.Distro = "aks-azurelinux-v2-gen2"
+			},
+			LiveVMValidators: []*LiveVMValidator{
+				SysctlConfigValidator(customSysctls),
+				UlimitValidator(customContainerdUlimits),
+			},
+		},
+	}
+}
+
+// Returns config for the 'gpu' E2E scenario
+func azurelinuxv2gpu() *Scenario {
+	return &Scenario{
+		Name:        "azurelinuxv2-gpu",
+		Description: "Tests that a GPU-enabled node using a AzureLinuxV2 (CgroupV2) VHD can be properly bootstrapped",
+		Tags: Tags{
+			Name:     "azurelinuxv2-gpu",
+			OS:       "azurelinuxv2",
+			Platform: "x64",
+			GPU:      true,
+		},
+		Config: Config{
+			Cluster:     cluster.ClusterKubenet,
+			VHDSelector: config.VHDAzureLinuxV2Gen2,
+			BootstrapConfigMutator: func(nbc *datamodel.NodeBootstrappingConfiguration) {
+				nbc.ContainerService.Properties.AgentPoolProfiles[0].VMSize = "Standard_NC6s_v3"
+				nbc.ContainerService.Properties.AgentPoolProfiles[0].Distro = "aks-azurelinux-v2-gen2"
+				nbc.AgentPoolProfile.VMSize = "Standard_NC6s_v3"
+				nbc.AgentPoolProfile.Distro = "aks-azurelinux-v2-gen2"
+				nbc.ConfigGPUDriverIfNeeded = true
+				nbc.EnableGPUDevicePluginIfNeeded = false
+				nbc.EnableNvidia = true
+			},
+			VMConfigMutator: func(vmss *armcompute.VirtualMachineScaleSet) {
+				vmss.SKU.Name = to.Ptr("Standard_NC6s_v3")
+			},
+		},
+	}
+}
+
+func azurelinuxv2gpu_azurecni() *Scenario {
+	return &Scenario{
+		Name:        "azurelinuxv2-gpu-azurecni",
+		Description: "AzureLinux V2 (CgroupV2) gpu scenario on cluster configured with Azure CNI",
+		Tags:        Tags{Name: "azurelinuxv2-gpu-azurecni", OS: "azurelinuxv2", Platform: "x64", GPU: true},
+		Config: Config{
+			Cluster:     cluster.ClusterAzureNetwork,
+			VHDSelector: config.VHDAzureLinuxV2Gen2,
+			BootstrapConfigMutator: func(nbc *datamodel.NodeBootstrappingConfiguration) {
+				nbc.ContainerService.Properties.OrchestratorProfile.KubernetesConfig.NetworkPlugin = string(armcontainerservice.NetworkPluginAzure)
+				nbc.AgentPoolProfile.KubernetesConfig.NetworkPlugin = string(armcontainerservice.NetworkPluginAzure)
+				nbc.ContainerService.Properties.AgentPoolProfiles[0].VMSize = "Standard_NC6s_v3"
+				nbc.ContainerService.Properties.AgentPoolProfiles[0].Distro = "aks-azurelinux-v2-gen2"
+				nbc.AgentPoolProfile.VMSize = "Standard_NC6s_v3"
+				nbc.AgentPoolProfile.Distro = "aks-azurelinux-v2-gen2"
+				nbc.ConfigGPUDriverIfNeeded = true
+				nbc.EnableGPUDevicePluginIfNeeded = false
+				nbc.EnableNvidia = true
+			},
+			VMConfigMutator: func(vmss *armcompute.VirtualMachineScaleSet) {
+				vmss.SKU.Name = to.Ptr("Standard_NC6s_v3")
+			},
+		},
+	}
+}
+
+func azurelinuxv2Wasm() *Scenario {
+	return &Scenario{
+		Name:        "azurelinuxv2-wasm",
+		Description: "tests that a new AzureLinuxV2 (CgroupV2) node using krustlet can be properly bootstrapped",
+		Tags: Tags{
+			Name:     "azurelinuxv2-wasm",
+			OS:       "azurelinuxv2",
+			Platform: "x64",
+			WASM:     true,
+		},
+		Config: Config{
+			Cluster:     cluster.ClusterKubenet,
+			VHDSelector: config.VHDAzureLinuxV2Gen2,
+			BootstrapConfigMutator: func(nbc *datamodel.NodeBootstrappingConfiguration) {
+				nbc.ContainerService.Properties.AgentPoolProfiles[0].WorkloadRuntime = datamodel.WasmWasi
+				nbc.ContainerService.Properties.AgentPoolProfiles[0].Distro = "aks-azurelinux-v2-gen2"
+				nbc.AgentPoolProfile.WorkloadRuntime = datamodel.WasmWasi
+				nbc.AgentPoolProfile.Distro = "aks-azurelinux-v2-gen2"
+			},
+		},
+	}
+}
+
+func marinerv2() *Scenario {
+	return &Scenario{
+		Name:        "marinerv2",
+		Description: "Tests that a node using a MarinerV2 VHD can be properly bootstrapped",
+		Tags: Tags{
+			Name:     "marinerv2",
+			OS:       "marinerv2",
+			Platform: "x64",
+		},
+		Config: Config{
+			Cluster:     cluster.ClusterKubenet,
+			VHDSelector: config.VHDCBLMarinerV2Gen2,
+			BootstrapConfigMutator: func(nbc *datamodel.NodeBootstrappingConfiguration) {
+				nbc.ContainerService.Properties.AgentPoolProfiles[0].Distro = "aks-cblmariner-v2-gen2"
+				nbc.AgentPoolProfile.Distro = "aks-cblmariner-v2-gen2"
+			},
+		},
+	}
+}
+
+func marinerv2AirGap() *Scenario {
+	return &Scenario{
+		Name:        "marinerv2-airgap",
+		Description: "Tests that a node using a MarinerV2 VHD can be properly bootstrapped",
+		Tags: Tags{
+			Name:     "marinerv2-airgap",
+			OS:       "marinerv2",
+			Platform: "x64",
+			Airgap:   true,
+		},
+		Config: Config{
+			Cluster:     cluster.ClusterKubenetAirgap,
+			VHDSelector: config.VHDCBLMarinerV2Gen2,
+			BootstrapConfigMutator: func(nbc *datamodel.NodeBootstrappingConfiguration) {
+				nbc.ContainerService.Properties.AgentPoolProfiles[0].Distro = "aks-cblmariner-v2-gen2"
+				nbc.AgentPoolProfile.Distro = "aks-cblmariner-v2-gen2"
+			},
+			Airgap: true,
+		},
+	}
+}
+
+func marinerv2ARM64() *Scenario {
+	return &Scenario{
+		Name:        "marinerv2-arm64",
+		Description: "Tests that a node using a MarinerV2 VHD on ARM64 architecture can be properly bootstrapped",
+		Tags: Tags{
+			Name:     "marinerv2-arm64",
+			OS:       "marinerv2",
+			Platform: "arm64",
+		},
+		Config: Config{
+			Cluster:     cluster.ClusterKubenet,
+			VHDSelector: config.VHDCBLMarinerV2Gen2Arm64,
+			BootstrapConfigMutator: func(nbc *datamodel.NodeBootstrappingConfiguration) {
+				nbc.ContainerService.Properties.AgentPoolProfiles[0].VMSize = "Standard_D2pds_V5"
+				nbc.ContainerService.Properties.AgentPoolProfiles[0].Distro = "aks-cblmariner-v2-arm64-gen2"
+				nbc.ContainerService.Properties.OrchestratorProfile.KubernetesConfig.CustomKubeBinaryURL = "https://acs-mirror.azureedge.net/kubernetes/v1.24.9/binaries/kubernetes-node-linux-arm64.tar.gz"
+				nbc.AgentPoolProfile.VMSize = "Standard_D2pds_V5"
+				nbc.AgentPoolProfile.Distro = "aks-cblmariner-v2-arm64-gen2"
+				nbc.IsARM64 = true
+			},
+			VMConfigMutator: func(vmss *armcompute.VirtualMachineScaleSet) {
+				vmss.SKU.Name = to.Ptr("Standard_D2pds_V5")
+			},
+		},
+	}
+}
+
+func marinerv2ARM64AirGap() *Scenario {
+	return &Scenario{
+		Name:        "marinerv2-arm64-airgap",
+		Description: "Tests that a node using a MarinerV2 VHD on ARM64 architecture can be properly bootstrapped",
+		Tags: Tags{
+			Name:     "marinerv2-arm64-airgap",
+			OS:       "marinerv2",
+			Platform: "arm64",
+			Airgap:   true,
+		},
+		Config: Config{
+			Cluster:     cluster.ClusterKubenetAirgap,
+			VHDSelector: config.VHDCBLMarinerV2Gen2Arm64,
+			BootstrapConfigMutator: func(nbc *datamodel.NodeBootstrappingConfiguration) {
+				nbc.ContainerService.Properties.AgentPoolProfiles[0].VMSize = "Standard_D2pds_V5"
+				nbc.ContainerService.Properties.AgentPoolProfiles[0].Distro = "aks-cblmariner-v2-arm64-gen2"
+				nbc.ContainerService.Properties.OrchestratorProfile.KubernetesConfig.CustomKubeBinaryURL = "https://acs-mirror.azureedge.net/kubernetes/v1.24.9/binaries/kubernetes-node-linux-arm64.tar.gz"
+				nbc.AgentPoolProfile.VMSize = "Standard_D2pds_V5"
+				nbc.AgentPoolProfile.Distro = "aks-cblmariner-v2-arm64-gen2"
+				nbc.IsARM64 = true
+			},
+			VMConfigMutator: func(vmss *armcompute.VirtualMachineScaleSet) {
+				vmss.SKU.Name = to.Ptr("Standard_D2pds_V5")
+			},
+			Airgap: true,
+		},
+	}
+}
+
+func marinerv2_azurecni() *Scenario {
+	return &Scenario{
+		Name:        "marinerv2-azurecni",
+		Description: "marinerv2 scenario on a cluster configured with Azure CNI",
+		Tags: Tags{
+			Name:     "marinerv2-azurecni",
+			OS:       "marinerv2",
+			Platform: "x64",
+		},
+		Config: Config{
+			Cluster:     cluster.ClusterAzureNetwork,
+			VHDSelector: config.VHDCBLMarinerV2Gen2,
+			BootstrapConfigMutator: func(nbc *datamodel.NodeBootstrappingConfiguration) {
+				nbc.ContainerService.Properties.OrchestratorProfile.KubernetesConfig.NetworkPlugin = string(armcontainerservice.NetworkPluginAzure)
+				nbc.AgentPoolProfile.KubernetesConfig.NetworkPlugin = string(armcontainerservice.NetworkPluginAzure)
+				nbc.ContainerService.Properties.AgentPoolProfiles[0].Distro = "aks-cblmariner-v2-gen2"
+				nbc.AgentPoolProfile.Distro = "aks-cblmariner-v2-gen2"
+			},
+		},
+	}
+}
+
+func marinerv2ChronyRestarts() *Scenario {
+	return &Scenario{
+		Name:        "marinerv2-chrony-restarts",
+		Description: "Tests that the chrony service restarts if it is killed",
+		Tags: Tags{
+			Name:     "marinerv2-chrony-restarts",
+			OS:       "marinerv2",
+			Platform: "x64",
+		},
+		Config: Config{
+			Cluster:     cluster.ClusterKubenet,
+			VHDSelector: config.VHDCBLMarinerV2Gen2,
+			BootstrapConfigMutator: func(nbc *datamodel.NodeBootstrappingConfiguration) {
+				nbc.ContainerService.Properties.AgentPoolProfiles[0].Distro = "aks-cblmariner-v2-gen2"
+				nbc.AgentPoolProfile.Distro = "aks-cblmariner-v2-gen2"
+			},
+			LiveVMValidators: []*LiveVMValidator{
+				serviceCanRestartValidator("chronyd", 10),
+				FileHasContentsValidator("/etc/systemd/system/chronyd.service.d/10-chrony-restarts.conf", "Restart=always"),
+				FileHasContentsValidator("/etc/systemd/system/chronyd.service.d/10-chrony-restarts.conf", "RestartSec=5"),
+			},
+		},
+	}
+}
+
+func marinerv2CustomSysctls() *Scenario {
+	customSysctls := map[string]string{
+		"net.ipv4.ip_local_port_range":       "32768 62535",
+		"net.netfilter.nf_conntrack_max":     "2097152",
+		"net.netfilter.nf_conntrack_buckets": "524288",
+		"net.ipv4.tcp_keepalive_intvl":       "90",
+		"net.ipv4.ip_local_reserved_ports":   "",
+	}
+	customContainerdUlimits := map[string]string{
+		"LimitMEMLOCK": "75000",
+		"LimitNOFILE":  "1048",
+	}
+	return &Scenario{
+		Name:        "marinerv2-custom-sysctls",
+		Description: "tests that a MarinerV2 VHD can be properly bootstrapped when supplied custom node config that contains custom sysctl settings",
+		Tags: Tags{
+			Name:     "marinerv2-custom-sysctls",
+			OS:       "marinerv2",
+			Platform: "x64",
+		},
+		Config: Config{
+			Cluster:     cluster.ClusterKubenet,
+			VHDSelector: config.VHDCBLMarinerV2Gen2,
+			BootstrapConfigMutator: func(nbc *datamodel.NodeBootstrappingConfiguration) {
+				customLinuxConfig := &datamodel.CustomLinuxOSConfig{
+					Sysctls: &datamodel.SysctlConfig{
+						NetNetfilterNfConntrackMax:     to.Ptr(toolkit.StrToInt32(customSysctls["net.netfilter.nf_conntrack_max"])),
+						NetNetfilterNfConntrackBuckets: to.Ptr(toolkit.StrToInt32(customSysctls["net.netfilter.nf_conntrack_buckets"])),
+						NetIpv4IpLocalPortRange:        customSysctls["net.ipv4.ip_local_port_range"],
+						NetIpv4TcpkeepaliveIntvl:       to.Ptr(toolkit.StrToInt32(customSysctls["net.ipv4.tcp_keepalive_intvl"])),
+					},
+					UlimitConfig: &datamodel.UlimitConfig{
+						MaxLockedMemory: customContainerdUlimits["LimitMEMLOCK"],
+						NoFile:          customContainerdUlimits["LimitNOFILE"],
+					},
+				}
+				nbc.AgentPoolProfile.CustomLinuxOSConfig = customLinuxConfig
+				nbc.ContainerService.Properties.AgentPoolProfiles[0].CustomLinuxOSConfig = customLinuxConfig
+				nbc.ContainerService.Properties.AgentPoolProfiles[0].Distro = "aks-cblmariner-v2-gen2"
+				nbc.AgentPoolProfile.Distro = "aks-cblmariner-v2-gen2"
+			},
+			LiveVMValidators: []*LiveVMValidator{
+				SysctlConfigValidator(customSysctls),
+				UlimitValidator(customContainerdUlimits),
+			},
+		},
+	}
+}
+
+// Returns config for the 'gpu' E2E scenario
+func marinerv2gpu() *Scenario {
+	return &Scenario{
+		Name:        "marinerv2-gpu",
+		Description: "Tests that a GPU-enabled node using a MarinerV2 VHD can be properly bootstrapped",
+		Tags: Tags{
+			Name:     "marinerv2-gpu",
+			OS:       "marinerv2",
+			Platform: "x64",
+			GPU:      true,
+		},
+		Config: Config{
+			Cluster:     cluster.ClusterKubenet,
+			VHDSelector: config.VHDCBLMarinerV2Gen2,
+			BootstrapConfigMutator: func(nbc *datamodel.NodeBootstrappingConfiguration) {
+				nbc.ContainerService.Properties.AgentPoolProfiles[0].VMSize = "Standard_NC6s_v3"
+				nbc.ContainerService.Properties.AgentPoolProfiles[0].Distro = "aks-cblmariner-v2-gen2"
+				nbc.AgentPoolProfile.VMSize = "Standard_NC6s_v3"
+				nbc.AgentPoolProfile.Distro = "aks-cblmariner-v2-gen2"
+				nbc.ConfigGPUDriverIfNeeded = true
+				nbc.EnableGPUDevicePluginIfNeeded = false
+				nbc.EnableNvidia = true
+			},
+			VMConfigMutator: func(vmss *armcompute.VirtualMachineScaleSet) {
+				vmss.SKU.Name = to.Ptr("Standard_NC6s_v3")
+			},
+		},
+	}
+}
+
+func marinerv2gpu_azurecni() *Scenario {
+	return &Scenario{
+		Name:        "marinerv2-gpu-azurecni",
+		Description: "MarinerV2 gpu scenario on cluster configured with Azure CNI",
+		Tags: Tags{
+			Name:     "marinerv2-gpu-azurecni",
+			OS:       "marinerv2",
+			Platform: "x64",
+			GPU:      true,
+		},
+		Config: Config{
+			Cluster:     cluster.ClusterAzureNetwork,
+			VHDSelector: config.VHDCBLMarinerV2Gen2,
+			BootstrapConfigMutator: func(nbc *datamodel.NodeBootstrappingConfiguration) {
+				nbc.ContainerService.Properties.OrchestratorProfile.KubernetesConfig.NetworkPlugin = string(armcontainerservice.NetworkPluginAzure)
+				nbc.AgentPoolProfile.KubernetesConfig.NetworkPlugin = string(armcontainerservice.NetworkPluginAzure)
+				nbc.ContainerService.Properties.AgentPoolProfiles[0].VMSize = "Standard_NC6s_v3"
+				nbc.ContainerService.Properties.AgentPoolProfiles[0].Distro = "aks-cblmariner-v2-gen2"
+				nbc.AgentPoolProfile.VMSize = "Standard_NC6s_v3"
+				nbc.AgentPoolProfile.Distro = "aks-cblmariner-v2-gen2"
+				nbc.ConfigGPUDriverIfNeeded = true
+				nbc.EnableGPUDevicePluginIfNeeded = false
+				nbc.EnableNvidia = true
+			},
+			VMConfigMutator: func(vmss *armcompute.VirtualMachineScaleSet) {
+				vmss.SKU.Name = to.Ptr("Standard_NC6s_v3")
+			},
+		},
+	}
+}
+
+func marinerv2Wasm() *Scenario {
+	return &Scenario{
+		Name:        "marinerv2-wasm",
+		Description: "tests that a new marinerv2 node using krustlet can be properly bootstrapped",
+		Tags: Tags{
+			Name:     "marinerv2-wasm",
+			OS:       "marinerv2",
+			Platform: "x64",
+			WASM:     true,
+		},
+		Config: Config{
+			Cluster:     cluster.ClusterKubenet,
+			VHDSelector: config.VHDCBLMarinerV2Gen2,
+			BootstrapConfigMutator: func(nbc *datamodel.NodeBootstrappingConfiguration) {
+				nbc.ContainerService.Properties.AgentPoolProfiles[0].WorkloadRuntime = datamodel.WasmWasi
+				nbc.ContainerService.Properties.AgentPoolProfiles[0].Distro = "aks-cblmariner-v2-gen2"
+				nbc.AgentPoolProfile.WorkloadRuntime = datamodel.WasmWasi
+				nbc.AgentPoolProfile.Distro = "aks-cblmariner-v2-gen2"
+			},
+		},
+	}
+}
+
+// Returns config for the 'base' E2E scenario
+func ubuntu1804() *Scenario {
+	return &Scenario{
+		Name:        "ubuntu1804",
+		Description: "Tests that a node using an Ubuntu 1804 VHD can be properly bootstrapped",
+		Tags: Tags{
+			Name:     "ubuntu1804",
+			OS:       "ubuntu1804",
+			Platform: "x64",
+		},
+		Config: Config{
+			Cluster:     cluster.ClusterKubenet,
+			VHDSelector: config.VHDUbuntu1804Gen2Containerd,
+		},
+	}
+}
+
+func ubuntu1804_azurecni() *Scenario {
+	return &Scenario{
+		Name:        "ubuntu1804-azurecni",
+		Description: "ubuntu1804 scenario on cluster configured with Azure CNI",
+		Tags: Tags{
+			Name:     "ubuntu1804-azurecni",
+			OS:       "ubuntu1804",
+			Platform: "x64",
+		},
+		Config: Config{
+			Cluster:     cluster.ClusterAzureNetwork,
+			VHDSelector: config.VHDUbuntu1804Gen2Containerd,
+			BootstrapConfigMutator: func(nbc *datamodel.NodeBootstrappingConfiguration) {
+				nbc.ContainerService.Properties.OrchestratorProfile.KubernetesConfig.NetworkPlugin = string(armcontainerservice.NetworkPluginAzure)
+				nbc.AgentPoolProfile.KubernetesConfig.NetworkPlugin = string(armcontainerservice.NetworkPluginAzure)
+			},
+		},
+	}
+}
+
+func ubuntu1804ChronyRestarts() *Scenario {
+	return &Scenario{
+		Name:        "ubuntu1804-chrony-restarts",
+		Description: "Tests that the chrony service restarts if it is killed",
+		Tags: Tags{
+			Name:     "ubuntu1804-chrony-restarts",
+			OS:       "ubuntu1804",
+			Platform: "x64",
+		},
+		Config: Config{
+			Cluster:     cluster.ClusterKubenet,
+			VHDSelector: config.VHDUbuntu1804Gen2Containerd,
+			BootstrapConfigMutator: func(nbc *datamodel.NodeBootstrappingConfiguration) {
+				nbc.ContainerService.Properties.AgentPoolProfiles[0].Distro = "aks-ubuntu-containerd-18.04-gen2"
+				nbc.AgentPoolProfile.Distro = "aks-ubuntu-containerd-18.04-gen2"
+			},
+			LiveVMValidators: []*LiveVMValidator{
+				serviceCanRestartValidator("chronyd", 10),
+				FileHasContentsValidator("/etc/systemd/system/chrony.service.d/10-chrony-restarts.conf", "Restart=always"),
+				FileHasContentsValidator("/etc/systemd/system/chrony.service.d/10-chrony-restarts.conf", "RestartSec=5"),
+			},
+		},
+	}
+}
+
+// Returns config for the 'gpu' E2E scenario
+func ubuntu1804gpu() *Scenario {
+	return &Scenario{
+		Name:        "ubuntu1804-gpu",
+		Description: "Tests that a GPU-enabled node using an Ubuntu 1804 VHD can be properly bootstrapped",
+		Tags: Tags{
+			Name:     "ubuntu1804-gpu",
+			OS:       "ubuntu1804",
+			Platform: "x64",
+			GPU:      true,
+		},
+		Config: Config{
+			Cluster:     cluster.ClusterKubenet,
+			VHDSelector: config.VHDUbuntu1804Gen2Containerd,
+			BootstrapConfigMutator: func(nbc *datamodel.NodeBootstrappingConfiguration) {
+				nbc.ContainerService.Properties.AgentPoolProfiles[0].VMSize = "Standard_NC6s_v3"
+				nbc.ContainerService.Properties.AgentPoolProfiles[0].Distro = "aks-ubuntu-containerd-18.04-gen2"
+				nbc.AgentPoolProfile.VMSize = "Standard_NC6s_v3"
+				nbc.AgentPoolProfile.Distro = "aks-ubuntu-containerd-18.04-gen2"
+				nbc.ConfigGPUDriverIfNeeded = true
+				nbc.EnableGPUDevicePluginIfNeeded = false
+				nbc.EnableNvidia = true
+			},
+			VMConfigMutator: func(vmss *armcompute.VirtualMachineScaleSet) {
+				vmss.SKU.Name = to.Ptr("Standard_NC6s_v3")
+			},
+		},
+	}
+}
+
+func ubuntu1804gpu_azurecni() *Scenario {
+	return &Scenario{
+		Name:        "ubuntu1804-gpu-azurecni",
+		Description: "Ubuntu1804 gpu scenario on cluster configured with Azure CNI",
+		Tags: Tags{
+			Name:     "ubuntu1804-gpu-azurecni",
+			OS:       "ubuntu1804",
+			Platform: "x64",
+			GPU:      true,
+		},
+		Config: Config{
+			Cluster:     cluster.ClusterAzureNetwork,
+			VHDSelector: config.VHDUbuntu1804Gen2Containerd,
+			BootstrapConfigMutator: func(nbc *datamodel.NodeBootstrappingConfiguration) {
+				nbc.ContainerService.Properties.OrchestratorProfile.KubernetesConfig.NetworkPlugin = string(armcontainerservice.NetworkPluginAzure)
+				nbc.AgentPoolProfile.KubernetesConfig.NetworkPlugin = string(armcontainerservice.NetworkPluginAzure)
+				nbc.ContainerService.Properties.AgentPoolProfiles[0].VMSize = "Standard_NC6s_v3"
+				nbc.ContainerService.Properties.AgentPoolProfiles[0].Distro = "aks-ubuntu-containerd-18.04-gen2"
+				nbc.AgentPoolProfile.VMSize = "Standard_NC6s_v3"
+				nbc.AgentPoolProfile.Distro = "aks-ubuntu-containerd-18.04-gen2"
+				nbc.ConfigGPUDriverIfNeeded = true
+				nbc.EnableGPUDevicePluginIfNeeded = false
+				nbc.EnableNvidia = true
+			},
+			VMConfigMutator: func(vmss *armcompute.VirtualMachineScaleSet) {
+				vmss.SKU.Name = to.Ptr("Standard_NC6s_v3")
+			},
+		},
+	}
+}
+
+func ubuntu2204() *Scenario {
+	return &Scenario{
+		Name:        "ubuntu2204",
+		Description: "Tests that a node using the Ubuntu 2204 VHD can be properly bootstrapped",
+		Tags: Tags{
+			Name:     "ubuntu2204",
+			OS:       "ubuntu2204",
+			Platform: "x64",
+		},
+		Config: Config{
+			Cluster:     cluster.ClusterKubenet,
+			VHDSelector: config.VHDUbuntu2204Gen2Containerd,
+			BootstrapConfigMutator: func(nbc *datamodel.NodeBootstrappingConfiguration) {
+				nbc.ContainerService.Properties.AgentPoolProfiles[0].Distro = "aks-ubuntu-containerd-22.04-gen2"
+				nbc.AgentPoolProfile.Distro = "aks-ubuntu-containerd-22.04-gen2"
+			},
+		},
+	}
+}
+
+func ubuntu2204AirGap() *Scenario {
+	return &Scenario{
+		Name:        "ubuntu2204-airgap",
+		Description: "Tests that a node using the Ubuntu 2204 VHD and is airgap can be properly bootstrapped",
+		Tags: Tags{
+			Name:     "ubuntu2204-airgap",
+			OS:       "ubuntu2204",
+			Platform: "x64",
+			Airgap:   true,
+		},
+		Config: Config{
+			Cluster:     cluster.ClusterKubenetAirgap,
+			VHDSelector: config.VHDUbuntu2204Gen2Containerd,
+			BootstrapConfigMutator: func(nbc *datamodel.NodeBootstrappingConfiguration) {
+				nbc.ContainerService.Properties.AgentPoolProfiles[0].Distro = "aks-ubuntu-containerd-22.04-gen2"
+				nbc.AgentPoolProfile.Distro = "aks-ubuntu-containerd-22.04-gen2"
+			},
+			Airgap: true,
+		},
+	}
+}
+
+func ubuntu2204ARM64() *Scenario {
+	return &Scenario{
+		Name:        "ubuntu2204-arm64",
+		Description: "Tests that an Ubuntu 2204 Node using ARM64 architecture can be properly bootstrapped",
+		Tags: Tags{
+			Name:     "ubuntu2204-arm64",
+			OS:       "ubuntu2204",
+			Platform: "arm64",
+		},
+		Config: Config{
+			Cluster:     cluster.ClusterKubenet,
+			VHDSelector: config.VHDUbuntu2204Gen2Arm64Containerd,
+			BootstrapConfigMutator: func(nbc *datamodel.NodeBootstrappingConfiguration) {
+				nbc.ContainerService.Properties.AgentPoolProfiles[0].VMSize = "Standard_D2pds_V5"
+				nbc.ContainerService.Properties.AgentPoolProfiles[0].Distro = "aks-ubuntu-arm64-containerd-22.04-gen2"
+				// This needs to be set based on current CSE implementation...
+				nbc.ContainerService.Properties.OrchestratorProfile.KubernetesConfig.CustomKubeBinaryURL = "https://acs-mirror.azureedge.net/kubernetes/v1.24.9/binaries/kubernetes-node-linux-arm64.tar.gz"
+				nbc.AgentPoolProfile.VMSize = "Standard_D2pds_V5"
+				nbc.AgentPoolProfile.Distro = "aks-ubuntu-arm64-containerd-22.04-gen2"
+				nbc.IsARM64 = true
+			},
+			VMConfigMutator: func(vmss *armcompute.VirtualMachineScaleSet) {
+				vmss.SKU.Name = to.Ptr("Standard_D2pds_V5")
+			},
+		},
+	}
+}
+
+func ubuntu2204ArtifactStreaming() *Scenario {
+	return &Scenario{
+		Name:        "ubuntu2204-artifact-streaming",
+		Description: "tests that a new ubuntu 2204 node using artifact streaming can be properly bootstrapepd",
+		Tags: Tags{
+			Name:     "ubuntu2204-artifact-streaming",
+			OS:       "ubuntu2204",
+			Platform: "x64",
+		},
+		Config: Config{
+			Cluster:     cluster.ClusterKubenet,
+			VHDSelector: config.VHDUbuntu2204Gen2Containerd,
+			BootstrapConfigMutator: func(nbc *datamodel.NodeBootstrappingConfiguration) {
+				nbc.EnableArtifactStreaming = true
+				nbc.ContainerService.Properties.AgentPoolProfiles[0].Distro = "aks-ubuntu-containerd-22.04-gen2"
+				nbc.AgentPoolProfile.Distro = "aks-ubuntu-containerd-22.04-gen2"
+			},
+			LiveVMValidators: []*LiveVMValidator{
+				NonEmptyDirectoryValidator("/etc/overlaybd"),
+			},
+		},
+	}
+}
+
+func ubuntu2204ChronyRestarts() *Scenario {
+	return &Scenario{
+		Name:        "ubuntu2204-chrony-restarts",
+		Description: "Tests that the chrony service restarts if it is killed",
+		Tags: Tags{
+			Name:     "ubuntu2204-chrony-restarts",
+			OS:       "ubuntu2204",
+			Platform: "x64",
+		},
+		Config: Config{
+			Cluster:     cluster.ClusterKubenet,
+			VHDSelector: config.VHDUbuntu2204Gen2Containerd,
+			BootstrapConfigMutator: func(nbc *datamodel.NodeBootstrappingConfiguration) {
+				nbc.ContainerService.Properties.AgentPoolProfiles[0].Distro = "aks-ubuntu-containerd-22.04-gen2"
+				nbc.AgentPoolProfile.Distro = "aks-ubuntu-containerd-22.04-gen2"
+			},
+			LiveVMValidators: []*LiveVMValidator{
+				serviceCanRestartValidator("chronyd", 10),
+				FileHasContentsValidator("/etc/systemd/system/chronyd.service.d/10-chrony-restarts.conf", "Restart=always"),
+				FileHasContentsValidator("/etc/systemd/system/chronyd.service.d/10-chrony-restarts.conf", "RestartSec=5"),
+			},
+		},
+	}
+}
+
+const encodedTestCert = "LS0tLS1CRUdJTiBDRVJUSUZJQ0FURS0tLS0tCk1JSUgvVENDQmVXZ0F3SUJBZ0lRYUJZRTMvTTA4WEhZQ25OVm1jRkJjakFOQmdrcWhraUc5dzBCQVFzRkFEQnkKTVFzd0NRWURWUVFHRXdKVlV6RU9NQXdHQTFVRUNBd0ZWR1Y0WVhNeEVEQU9CZ05WQkFjTUIwaHZkWE4wYjI0eApFVEFQQmdOVkJBb01DRk5UVENCRGIzSndNUzR3TEFZRFZRUUREQ1ZUVTB3dVkyOXRJRVZXSUZOVFRDQkpiblJsCmNtMWxaR2xoZEdVZ1EwRWdVbE5CSUZJek1CNFhEVEl3TURRd01UQXdOVGd6TTFvWERUSXhNRGN4TmpBd05UZ3oKTTFvd2diMHhDekFKQmdOVkJBWVRBbFZUTVE0d0RBWURWUVFJREFWVVpYaGhjekVRTUE0R0ExVUVCd3dIU0c5MQpjM1J2YmpFUk1BOEdBMVVFQ2d3SVUxTk1JRU52Y25BeEZqQVVCZ05WQkFVVERVNVdNakF3T0RFMk1UUXlORE14CkZEQVNCZ05WQkFNTUMzZDNkeTV6YzJ3dVkyOXRNUjB3R3dZRFZRUVBEQlJRY21sMllYUmxJRTl5WjJGdWFYcGgKZEdsdmJqRVhNQlVHQ3lzR0FRUUJnamM4QWdFQ0RBWk9aWFpoWkdFeEV6QVJCZ3NyQmdFRUFZSTNQQUlCQXhNQwpWVk13Z2dFaU1BMEdDU3FHU0liM0RRRUJBUVVBQTRJQkR3QXdnZ0VLQW9JQkFRREhoZVJrYmIxRkNjN3hSS3N0CndLMEpJR2FLWTh0N0piUzJiUTJiNllJSkRnbkh1SVlIcUJyQ1VWNzlvZWxpa2tva1JrRnZjdnBhS2luRkhEUUgKVXBXRUk2UlVFUlltU0NnM084V2k0MnVPY1YyQjVaYWJtWENrd2R4WTVFY2w1MUJiTThVbkdkb0FHYmRObWlSbQpTbVRqY3MrbGhNeGc0ZkZZNmxCcGlFVkZpR1VqR1JSKzYxUjY3THo2VTRLSmVMTmNDbTA3UXdGWUtCbXBpMDhnCmR5Z1N2UmRVdzU1Sm9wcmVkaitWR3RqVWtCNGhGVDRHUVgvZ2h0NjlSbHF6Lys4dTBkRVFraHVVdXVjcnFhbG0KU0d5NDNIUndCZkRLRndZZVdNN0NQTWQ1ZS9kTyt0MDh0OFBianpWVFR2NWhRRENzRVlJVjJUN0FGSTlTY054TQpraDcvQWdNQkFBR2pnZ05CTUlJRFBUQWZCZ05WSFNNRUdEQVdnQlMvd1ZxSC95ajZRVDM5dDAva0hhK2dZVmdwCnZUQi9CZ2dyQmdFRkJRY0JBUVJ6TUhFd1RRWUlLd1lCQlFVSE1BS0dRV2gwZEhBNkx5OTNkM2N1YzNOc0xtTnYKYlM5eVpYQnZjMmwwYjNKNUwxTlRUR052YlMxVGRXSkRRUzFGVmkxVFUwd3RVbE5CTFRRd09UWXRVak11WTNKMApNQ0FHQ0NzR0FRVUZCekFCaGhSb2RIUndPaTh2YjJOemNITXVjM05zTG1OdmJUQWZCZ05WSFJFRUdEQVdnZ3QzCmQzY3VjM05zTG1OdmJZSUhjM05zTG1OdmJUQmZCZ05WSFNBRVdEQldNQWNHQldlQkRBRUJNQTBHQ3lxRWFBR0cKOW5jQ0JRRUJNRHdHRENzR0FRUUJncWt3QVFNQkJEQXNNQ29HQ0NzR0FRVUZCd0lCRmg1b2RIUndjem92TDNkMwpkeTV6YzJ3dVkyOXRMM0psY0c5emFYUnZjbmt3SFFZRFZSMGxCQll3RkFZSUt3WUJCUVVIQXdJR0NDc0dBUVVGCkJ3TUJNRWdHQTFVZEh3UkJNRDh3UGFBN29EbUdOMmgwZEhBNkx5OWpjbXh6TG5OemJDNWpiMjB2VTFOTVkyOXQKTFZOMVlrTkJMVVZXTFZOVFRDMVNVMEV0TkRBNU5pMVNNeTVqY213d0hRWURWUjBPQkJZRUZBREFGVUlhenc1cgpaSUhhcG5SeElVbnB3K0dMTUE0R0ExVWREd0VCL3dRRUF3SUZvRENDQVgwR0Npc0dBUVFCMW5rQ0JBSUVnZ0Z0CkJJSUJhUUZuQUhjQTlseVVMOUYzTUNJVVZCZ0lNSlJXanVOTkV4a3p2OThNTHlBTHpFN3haT01BQUFGeE0waG8KYndBQUJBTUFTREJHQWlFQTZ4ZWxpTlI4R2svNjNwWWRuUy92T3gvQ2pwdEVNRXY4OVdXaDEvdXJXSUVDSVFEeQpCcmVIVTI1RHp3dWtRYVJRandXNjU1WkxrcUNueGJ4UVdSaU9lbWo5SkFCMUFKUWd2QjZPMVkxc2lITWZnb3NpCkxBM1IyazFlYkUrVVBXSGJUaTlZVGFMQ0FBQUJjVE5JYU53QUFBUURBRVl3UkFJZ0dSRTR3emFiTlJkRDhrcS8KdkZQM3RRZTJobTB4NW5YdWxvd2g0SWJ3M2xrQ0lGWWIvM2xTRHBsUzdBY1I0citYcFd0RUtTVEZXSm1OQ1JiYwpYSnVyMlJHQkFIVUE3c0NWN28xeVpBK1M0OE81RzhjU28ybHFDWHRMYWhvVU9PWkhzc3Z0eGZrQUFBRnhNMGhvCjh3QUFCQU1BUmpCRUFpQjZJdmJvV3NzM1I0SXRWd2plYmw3RDN5b0ZhWDBORGgyZFdoaGd3Q3hySHdJZ0NmcTcKb2NNQzV0KzFqaTVNNXhhTG1QQzRJK1dYM0kvQVJrV1N5aU83SVFjd0RRWUpLb1pJaHZjTkFRRUxCUUFEZ2dJQgpBQ2V1dXI0UW51anFtZ3VTckhVM21oZitjSm9kelRRTnFvNHRkZStQRDEvZUZkWUFFTHU4eEYrMEF0N3hKaVBZCmk1Ukt3aWx5UDU2diszaVkyVDlsdzdTOFRKMDQxVkxoYUlLcDE0TXpTVXpSeWVvT0FzSjdRQURNQ2xIS1VEbEgKVVUycE51bzg4WTZpZ292VDNic253Sk5pRVFOcXltU1NZaGt0dzB0YWR1b3FqcVhuMDZnc1Zpb1dUVkRYeXNkNQpxRXg0dDZzSWdJY01tMjZZSDF2SnBDUUVoS3BjMnkwN2dSa2tsQlpSdE1qVGh2NGNYeXlNWDd1VGNkVDdBSkJQCnVlaWZDb1YyNUp4WHVvOGQ1MTM5Z3dQMUJBZTdJQlZQeDJ1N0tOL1V5T1hkWm13TWYvVG1GR3dEZENmc3lIZi8KWnNCMndMSG96VFlvQVZtUTlGb1UxSkxnY1ZpdnFKK3ZObEJoSFhobHhNZE4wajgwUjlOejZFSWdsUWplSzNPOApJL2NGR20vQjgrNDJoT2xDSWQ5WmR0bmRKY1JKVmppMHdEMHF3ZXZDYWZBOWpKbEh2L2pzRStJOVV6NmNwQ3loCnN3K2xyRmR4VWdxVTU4YXhxZUs4OUZSK05vNHEwSUlPK0ppMXJKS3I5bmtTQjBCcVhvelZuRTFZQi9LTHZkSXMKdVlaSnVxYjJwS2t1K3p6VDZnVXdIVVRadkJpTk90WEw0Tnh3Yy9LVDdXek9TZDJ3UDEwUUk4REtnNHZmaU5EcwpIV21CMWM0S2ppNmdPZ0E1dVNVemFHbXEvdjRWbmNLNVVyK245TGJmbmZMYzI4SjVmdC9Hb3Rpbk15RGszaWFyCkYxMFlscWNPbWVYMXVGbUtiZGkvWG9yR2xrQ29NRjNURHg4cm1wOURCaUIvCi0tLS0tRU5EIENFUlRJRklDQVRFLS0tLS0=" //nolint:lll
+
+func ubuntu2204CustomCATrust() *Scenario {
+	return &Scenario{
+		Name:        "ubuntu2204-custom-ca-trust",
+		Description: "Tests that a node using the Ubuntu 2204 VHD can be properly bootstrapped and custom CA was correctly added",
+		Tags: Tags{
+			Name:     "ubuntu2204-custom-ca-trust",
+			OS:       "ubuntu2204",
+			Platform: "x64",
+		},
+		Config: Config{
+			Cluster:     cluster.ClusterKubenet,
+			VHDSelector: config.VHDUbuntu2204Gen2Containerd,
+			BootstrapConfigMutator: func(nbc *datamodel.NodeBootstrappingConfiguration) {
+				nbc.ContainerService.Properties.AgentPoolProfiles[0].Distro = "aks-ubuntu-containerd-22.04-gen2"
+				nbc.AgentPoolProfile.Distro = "aks-ubuntu-containerd-22.04-gen2"
+				nbc.CustomCATrustConfig = &datamodel.CustomCATrustConfig{
+					CustomCATrustCerts: []string{
+						encodedTestCert,
+					},
+				}
+			},
+			LiveVMValidators: []*LiveVMValidator{
+				NonEmptyDirectoryValidator("/usr/local/share/ca-certificates/certs"),
+			},
+		},
+	}
+}
+
+func ubuntu2204CustomSysctls() *Scenario {
+	customSysctls := map[string]string{
+		"net.ipv4.ip_local_port_range":       "32768 65535",
+		"net.netfilter.nf_conntrack_max":     "2097152",
+		"net.netfilter.nf_conntrack_buckets": "524288",
+		"net.ipv4.tcp_keepalive_intvl":       "90",
+		"net.ipv4.ip_local_reserved_ports":   "65330",
+	}
+	customContainerdUlimits := map[string]string{
+		"LimitMEMLOCK": "75000",
+		"LimitNOFILE":  "1048",
+	}
+	return &Scenario{
+		Name:        "ubuntu2204-custom-sysctls",
+		Description: "tests that an ubuntu 2204 VHD can be properly bootstrapped when supplied custom node config that contains custom sysctl settings",
+		Tags: Tags{
+			Name:     "ubuntu2204-custom-sysctls",
+			OS:       "ubuntu2204",
+			Platform: "x64",
+		},
+		Config: Config{
+			Cluster:     cluster.ClusterKubenet,
+			VHDSelector: config.VHDUbuntu2204Gen2Containerd,
+			BootstrapConfigMutator: func(nbc *datamodel.NodeBootstrappingConfiguration) {
+				customLinuxConfig := &datamodel.CustomLinuxOSConfig{
+					Sysctls: &datamodel.SysctlConfig{
+						NetNetfilterNfConntrackMax:     to.Ptr(toolkit.StrToInt32(customSysctls["net.netfilter.nf_conntrack_max"])),
+						NetNetfilterNfConntrackBuckets: to.Ptr(toolkit.StrToInt32(customSysctls["net.netfilter.nf_conntrack_buckets"])),
+						NetIpv4IpLocalPortRange:        customSysctls["net.ipv4.ip_local_port_range"],
+						NetIpv4TcpkeepaliveIntvl:       to.Ptr(toolkit.StrToInt32(customSysctls["net.ipv4.tcp_keepalive_intvl"])),
+					},
+					UlimitConfig: &datamodel.UlimitConfig{
+						MaxLockedMemory: "75000",
+						NoFile:          "1048",
+					},
+				}
+				nbc.AgentPoolProfile.CustomLinuxOSConfig = customLinuxConfig
+				nbc.ContainerService.Properties.AgentPoolProfiles[0].CustomLinuxOSConfig = customLinuxConfig
+				nbc.ContainerService.Properties.AgentPoolProfiles[0].Distro = "aks-ubuntu-containerd-22.04-gen2"
+				nbc.AgentPoolProfile.Distro = "aks-ubuntu-containerd-22.04-gen2"
+			},
+			LiveVMValidators: []*LiveVMValidator{
+				SysctlConfigValidator(customSysctls),
+				UlimitValidator(customContainerdUlimits),
+			},
+		},
+	}
+}
+
+// Returns config for the 'gpu' E2E scenario
+func ubuntu2204gpu(name string, vmSize string) *Scenario {
+	return &Scenario{
+		Name:        name,
+		Description: fmt.Sprintf("Tests that a GPU-enabled node with VM size %s using an Ubuntu 2204 VHD can be properly bootstrapped", vmSize),
+		Tags:        Tags{Name: name, OS: "ubuntu2204", Platform: "x64", GPU: true},
+		Config: Config{
+			Cluster:     cluster.ClusterKubenet,
+			VHDSelector: config.VHDUbuntu2204Gen2Containerd,
+			BootstrapConfigMutator: func(nbc *datamodel.NodeBootstrappingConfiguration) {
+				nbc.ContainerService.Properties.AgentPoolProfiles[0].VMSize = vmSize
+				nbc.ContainerService.Properties.AgentPoolProfiles[0].Distro = "aks-ubuntu-containerd-22.04-gen2"
+				nbc.AgentPoolProfile.VMSize = vmSize
+				nbc.AgentPoolProfile.Distro = "aks-ubuntu-containerd-22.04-gen2"
+				nbc.ConfigGPUDriverIfNeeded = true
+				nbc.EnableGPUDevicePluginIfNeeded = false
+				nbc.EnableNvidia = true
+			},
+			VMConfigMutator: func(vmss *armcompute.VirtualMachineScaleSet) {
+				vmss.SKU.Name = to.Ptr(vmSize)
+			},
+		},
+	}
+}
+
+func ubuntu2204GPUGridDriver() *Scenario {
+	return &Scenario{
+		Name:        "ubuntu2204-gpu-grid",
+		Description: "Tests that a GPU-enabled node using the Ubuntu 2204 VHD with grid driver can be properly bootstrapped",
+		Tags: Tags{
+			Name:     "ubuntu2204-gpu-grid",
+			OS:       "ubuntu2204",
+			Platform: "x64",
+			GPU:      true,
+		},
+		Config: Config{
+			Cluster:     cluster.ClusterKubenet,
+			VHDSelector: config.VHDUbuntu2204Gen2Containerd,
+			BootstrapConfigMutator: func(nbc *datamodel.NodeBootstrappingConfiguration) {
+				nbc.ContainerService.Properties.AgentPoolProfiles[0].Distro = "aks-ubuntu-containerd-22.04-gen2"
+				nbc.AgentPoolProfile.Distro = "aks-ubuntu-containerd-22.04-gen2"
+				nbc.AgentPoolProfile.VMSize = "Standard_NV6ads_A10_v5"
+				nbc.ConfigGPUDriverIfNeeded = true
+				nbc.EnableGPUDevicePluginIfNeeded = false
+				nbc.EnableNvidia = true
+			},
+			VMConfigMutator: func(vmss *armcompute.VirtualMachineScaleSet) {
+				vmss.SKU.Name = to.Ptr("Standard_NV6ads_A10_v5")
+			},
+			LiveVMValidators: []*LiveVMValidator{
+				NvidiaSMIInstalledValidator(),
+			},
+		},
+	}
+}
+
+func ubuntu2204gpuNoDriver() *Scenario {
+	return &Scenario{
+		Name:        "ubuntu2204-gpu-nodriver",
+		Description: "Tests that a GPU-enabled node using the Ubuntu 2204 VHD opting for skipping gpu driver installation can be properly bootstrapped",
+		Tags: Tags{
+			Name:     "ubuntu2204-gpu-nodriver",
+			OS:       "ubuntu2204",
+			Platform: "x64",
+			GPU:      true,
+		},
+		Config: Config{
+			Cluster:     cluster.ClusterKubenet,
+			VHDSelector: config.VHDUbuntu2204Gen2Containerd,
+			BootstrapConfigMutator: func(nbc *datamodel.NodeBootstrappingConfiguration) {
+				nbc.ContainerService.Properties.AgentPoolProfiles[0].Distro = "aks-ubuntu-containerd-22.04-gen2"
+				nbc.AgentPoolProfile.Distro = "aks-ubuntu-containerd-22.04-gen2"
+				nbc.AgentPoolProfile.VMSize = "Standard_NC6s_v3"
+				nbc.ConfigGPUDriverIfNeeded = true
+				nbc.EnableGPUDevicePluginIfNeeded = false
+				nbc.EnableNvidia = true
+			},
+			VMConfigMutator: func(vmss *armcompute.VirtualMachineScaleSet) {
+				vmss.Tags = map[string]*string{
+					// deliberately case mismatched to agentbaker logic to check case insensitivity
+					"SkipGPUDriverInstall": to.Ptr("true"),
+				}
+				vmss.SKU.Name = to.Ptr("Standard_NC6s_v3")
+			},
+			LiveVMValidators: []*LiveVMValidator{
+				NvidiaSMINotInstalledValidator(),
+			},
+		},
+	}
+}
+
+func ubuntu2204privatekubepkg() *Scenario {
+	return &Scenario{
+		Name:        "ubuntu2204privatekubepkg",
+		Description: "Tests that a node using the Ubuntu 2204 VHD that was built with private kube packages can be properly bootstrapped with the specified kube version",
+		Tags: Tags{
+			Name:     "ubuntu2204privatekubepkg",
+			OS:       "ubuntu2204",
+			Platform: "x64",
+		},
+		Config: Config{
+			Cluster:     cluster.ClusterKubenet,
+			VHDSelector: config.VHDUbuntu2204Gen2ContainerdPrivateKubePkg,
+			BootstrapConfigMutator: func(nbc *datamodel.NodeBootstrappingConfiguration) {
+				nbc.ContainerService.Properties.AgentPoolProfiles[0].Distro = "aks-ubuntu-containerd-22.04-gen2"
+				nbc.ContainerService.Properties.OrchestratorProfile.OrchestratorVersion = "1.25.6"
+				nbc.AgentPoolProfile.Distro = "aks-ubuntu-containerd-22.04-gen2"
+				nbc.K8sComponents.LinuxPrivatePackageURL = "https://privatekube.blob.core.windows.net/kubernetes/v1.25.6-hotfix.20230612/binaries/v1.25.6-hotfix.20230612.tar.gz"
+			},
+		},
+	}
+}
+
+// These tests were created to verify that the apt-get call in downloadContainerdFromVersion is not executed.
+// The code path is not hit in either of these tests. In the future, testing with some kind of firewall to ensure no egress
+// calls are made would be beneficial for airgap testing.
+
+func ubuntu2204ContainerdURL() *Scenario {
+	return &Scenario{
+		Name:        "ubuntu2204ContainerdURL",
+		Description: "tests that a node using the Ubuntu 2204 VHD with the ContainerdPackageURL override bootstraps with the provided URL and not the maifest contianerd version",
+		Tags: Tags{
+			Name:     "ubuntu2204ContainerdURL",
+			OS:       "ubuntu2204",
+			Platform: "x64",
+		},
+		Config: Config{
+			Cluster:     cluster.ClusterKubenet,
+			VHDSelector: config.VHDUbuntu2204Gen2Containerd,
+			BootstrapConfigMutator: func(nbc *datamodel.NodeBootstrappingConfiguration) {
+				nbc.ContainerService.Properties.AgentPoolProfiles[0].Distro = "aks-ubuntu-containerd-22.04-gen2"
+				nbc.AgentPoolProfile.Distro = "aks-ubuntu-containerd-22.04-gen2"
+				nbc.ContainerdPackageURL = "https://packages.microsoft.com/ubuntu/22.04/prod/pool/main/m/moby-containerd/moby-containerd_1.6.9+azure-ubuntu22.04u1_amd64.deb"
+			},
+			LiveVMValidators: []*LiveVMValidator{
+				containerdVersionValidator("1.6.9"),
+			},
+		},
+	}
+}
+
+func ubuntu2204ContainerdVersion() *Scenario {
+	return &Scenario{
+		Name:        "ubuntu2204ContainerdVersion",
+		Description: "tests that a node using an Ubuntu2204 VHD and the ContainerdVersion override bootstraps with the correct manifest containerd version and ignores the override",
+		Config: Config{
+			Cluster:     cluster.ClusterKubenet,
+			VHDSelector: config.VHDUbuntu2204Gen2Containerd,
+			BootstrapConfigMutator: func(nbc *datamodel.NodeBootstrappingConfiguration) {
+				nbc.ContainerService.Properties.AgentPoolProfiles[0].Distro = "aks-ubuntu-containerd-22.04-gen2"
+				nbc.AgentPoolProfile.Distro = "aks-ubuntu-containerd-22.04-gen2"
+				nbc.ContainerdVersion = "1.6.9"
+			},
+			LiveVMValidators: []*LiveVMValidator{
+				containerdVersionValidator(getContainerdManifestVersion()),
+			},
+		},
+	}
+}
+
+func getContainerdManifestVersion() string {
+	manifest, err := getVHDManifest()
+	if err != nil {
+		panic(err)
+	}
+
+	return manifest.Containerd.Edge
+}
+
+func ubuntu2204Wasm() *Scenario {
+	return &Scenario{
+		Name:        "ubuntu2204-wasm",
+		Description: "tests that a new ubuntu 2204 node using krustlet can be properly bootstrapepd",
+		Tags: Tags{
+			Name:     "ubuntu2204-wasm",
+			OS:       "ubuntu2204",
+			Platform: "x64",
+		},
+		Config: Config{
+			Cluster:     cluster.ClusterKubenet,
+			VHDSelector: config.VHDUbuntu2204Gen2Containerd,
+			BootstrapConfigMutator: func(nbc *datamodel.NodeBootstrappingConfiguration) {
+				nbc.ContainerService.Properties.AgentPoolProfiles[0].WorkloadRuntime = datamodel.WasmWasi
+				nbc.ContainerService.Properties.AgentPoolProfiles[0].Distro = "aks-ubuntu-containerd-22.04-gen2"
+				nbc.AgentPoolProfile.WorkloadRuntime = datamodel.WasmWasi
+				nbc.AgentPoolProfile.Distro = "aks-ubuntu-containerd-22.04-gen2"
+			},
+		},
+	}
+}
