--- e2e/scenario/scenario_azurelinuxv2-airgap.go
@@ -1,29 +1,0 @@
-package scenario
-
-import (
-	"github.com/Azure/agentbaker/pkg/agent/datamodel"
-	"github.com/Azure/agentbakere2e/cluster"
-	"github.com/Azure/agentbakere2e/config"
-)
-
-func azurelinuxv2AirGap() *Scenario {
-	return &Scenario{
-		Name:        "azurelinuxv2-airgap",
-		Description: "Tests that a node using a AzureLinuxV2 (CgroupV2) VHD can be properly bootstrapped",
-		Tags: Tags{
-			Name:     "azurelinuxv2-airgap",
-			OS:       "azurelinuxv2",
-			Platform: "x64",
-			Airgap:   true,
-		},
-		Config: Config{
-			Cluster:     cluster.ClusterKubenetAirgap,
-			VHDSelector: config.VHDAzureLinuxV2Gen2,
-			BootstrapConfigMutator: func(nbc *datamodel.NodeBootstrappingConfiguration) {
-				nbc.ContainerService.Properties.AgentPoolProfiles[0].Distro = "aks-azurelinux-v2-gen2"
-				nbc.AgentPoolProfile.Distro = "aks-azurelinux-v2-gen2"
-			},
-			Airgap: true,
-		},
-	}
-}
