--- e2e/scenario/scenario_azurelinuxv2-azurecni.go
@@ -1,30 +1,0 @@
-package scenario
-
-import (
-	"github.com/Azure/agentbaker/pkg/agent/datamodel"
-	"github.com/Azure/agentbakere2e/cluster"
-	"github.com/Azure/agentbakere2e/config"
-	"github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/containerservice/armcontainerservice"
-)
-
-func azurelinuxv2_azurecni() *Scenario {
-	return &Scenario{
-		Name:        "azurelinuxv2-azurecni",
-		Description: "azurelinuxv2 scenario on a cluster configured with Azure CNI",
-		Tags: Tags{
-			Name:     "azurelinuxv2-azurecni",
-			OS:       "azurelinuxv2",
-			Platform: "x64",
-		},
-		Config: Config{
-			Cluster:     cluster.ClusterAzureNetwork,
-			VHDSelector: config.VHDAzureLinuxV2Gen2,
-			BootstrapConfigMutator: func(nbc *datamodel.NodeBootstrappingConfiguration) {
-				nbc.ContainerService.Properties.OrchestratorProfile.KubernetesConfig.NetworkPlugin = string(armcontainerservice.NetworkPluginAzure)
-				nbc.AgentPoolProfile.KubernetesConfig.NetworkPlugin = string(armcontainerservice.NetworkPluginAzure)
-				nbc.ContainerService.Properties.AgentPoolProfiles[0].Distro = "aks-azurelinux-v2-gen2"
-				nbc.AgentPoolProfile.Distro = "aks-azurelinux-v2-gen2"
-			},
-		},
-	}
-}
