--- e2e/scenario/scenario_azurelinuxv2-custom-sysctls.go
@@ -1,58 +1,0 @@
-package scenario
-
-import (
-	"github.com/Azure/agentbaker/pkg/agent/datamodel"
-	"github.com/Azure/agentbakere2e/cluster"
-	"github.com/Azure/agentbakere2e/config"
-	"github.com/Azure/agentbakere2e/toolkit"
-	"github.com/Azure/azure-sdk-for-go/sdk/azcore/to"
-)
-
-func azurelinuxv2CustomSysctls() *Scenario {
-	customSysctls := map[string]string{
-		"net.ipv4.ip_local_port_range":       "32768 62535",
-		"net.netfilter.nf_conntrack_max":     "2097152",
-		"net.netfilter.nf_conntrack_buckets": "524288",
-		"net.ipv4.tcp_keepalive_intvl":       "90",
-		"net.ipv4.ip_local_reserved_ports":   "",
-	}
-	customContainerdUlimits := map[string]string{
-		"LimitMEMLOCK": "75000",
-		"LimitNOFILE":  "1048",
-	}
-	return &Scenario{
-		Name:        "azurelinuxv2-custom-sysctls",
-		Description: "tests that a AzureLinuxV2 (CgroupV2) VHD can be properly bootstrapped when supplied custom node config that contains custom sysctl settings",
-		Tags: Tags{
-			Name:     "azurelinuxv2-custom-sysctls",
-			OS:       "azurelinuxv2",
-			Platform: "x64",
-		},
-		Config: Config{
-			Cluster:     cluster.ClusterKubenet,
-			VHDSelector: config.VHDAzureLinuxV2Gen2,
-			BootstrapConfigMutator: func(nbc *datamodel.NodeBootstrappingConfiguration) {
-				customLinuxConfig := &datamodel.CustomLinuxOSConfig{
-					Sysctls: &datamodel.SysctlConfig{
-						NetNetfilterNfConntrackMax:     to.Ptr(toolkit.StrToInt32(customSysctls["net.netfilter.nf_conntrack_max"])),
-						NetNetfilterNfConntrackBuckets: to.Ptr(toolkit.StrToInt32(customSysctls["net.netfilter.nf_conntrack_buckets"])),
-						NetIpv4IpLocalPortRange:        customSysctls["net.ipv4.ip_local_port_range"],
-						NetIpv4TcpkeepaliveIntvl:       to.Ptr(toolkit.StrToInt32(customSysctls["net.ipv4.tcp_keepalive_intvl"])),
-					},
-					UlimitConfig: &datamodel.UlimitConfig{
-						MaxLockedMemory: customContainerdUlimits["LimitMEMLOCK"],
-						NoFile:          customContainerdUlimits["LimitNOFILE"],
-					},
-				}
-				nbc.AgentPoolProfile.CustomLinuxOSConfig = customLinuxConfig
-				nbc.ContainerService.Properties.AgentPoolProfiles[0].CustomLinuxOSConfig = customLinuxConfig
-				nbc.ContainerService.Properties.AgentPoolProfiles[0].Distro = "aks-azurelinux-v2-gen2"
-				nbc.AgentPoolProfile.Distro = "aks-azurelinux-v2-gen2"
-			},
-			LiveVMValidators: []*LiveVMValidator{
-				SysctlConfigValidator(customSysctls),
-				UlimitValidator(customContainerdUlimits),
-			},
-		},
-	}
-}
