--- e2e/scenario/scenario_ubuntu1804-gpu-azurecni.go
@@ -1,41 +1,0 @@
-package scenario
-
-import (
-	"github.com/Azure/agentbaker/pkg/agent/datamodel"
-	"github.com/Azure/agentbakere2e/cluster"
-	"github.com/Azure/agentbakere2e/config"
-	"github.com/Azure/azure-sdk-for-go/sdk/azcore/to"
-	"github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/compute/armcompute"
-	"github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/containerservice/armcontainerservice"
-)
-
-func ubuntu1804gpu_azurecni() *Scenario {
-	return &Scenario{
-		Name:        "ubuntu1804-gpu-azurecni",
-		Description: "Ubuntu1804 gpu scenario on cluster configured with Azure CNI",
-		Tags: Tags{
-			Name:     "ubuntu1804-gpu-azurecni",
-			OS:       "ubuntu1804",
-			Platform: "x64",
-			GPU:      true,
-		},
-		Config: Config{
-			Cluster:     cluster.ClusterAzureNetwork,
-			VHDSelector: config.VHDUbuntu1804Gen2Containerd,
-			BootstrapConfigMutator: func(nbc *datamodel.NodeBootstrappingConfiguration) {
-				nbc.ContainerService.Properties.OrchestratorProfile.KubernetesConfig.NetworkPlugin = string(armcontainerservice.NetworkPluginAzure)
-				nbc.AgentPoolProfile.KubernetesConfig.NetworkPlugin = string(armcontainerservice.NetworkPluginAzure)
-				nbc.ContainerService.Properties.AgentPoolProfiles[0].VMSize = "Standard_NC6s_v3"
-				nbc.ContainerService.Properties.AgentPoolProfiles[0].Distro = "aks-ubuntu-containerd-18.04-gen2"
-				nbc.AgentPoolProfile.VMSize = "Standard_NC6s_v3"
-				nbc.AgentPoolProfile.Distro = "aks-ubuntu-containerd-18.04-gen2"
-				nbc.ConfigGPUDriverIfNeeded = true
-				nbc.EnableGPUDevicePluginIfNeeded = false
-				nbc.EnableNvidia = true
-			},
-			VMConfigMutator: func(vmss *armcompute.VirtualMachineScaleSet) {
-				vmss.SKU.Name = to.Ptr("Standard_NC6s_v3")
-			},
-		},
-	}
-}
