--- e2e/scenario/scenario_ubuntu2204-airgap.go
@@ -1,29 +1,0 @@
-package scenario
-
-import (
-	"github.com/Azure/agentbaker/pkg/agent/datamodel"
-	"github.com/Azure/agentbakere2e/cluster"
-	"github.com/Azure/agentbakere2e/config"
-)
-
-func ubuntu2204AirGap() *Scenario {
-	return &Scenario{
-		Name:        "ubuntu2204-airgap",
-		Description: "Tests that a node using the Ubuntu 2204 VHD and is airgap can be properly bootstrapped",
-		Tags: Tags{
-			Name:     "ubuntu2204-airgap",
-			OS:       "ubuntu2204",
-			Platform: "x64",
-			Airgap:   true,
-		},
-		Config: Config{
-			Cluster:     cluster.ClusterKubenetAirgap,
-			VHDSelector: config.VHDUbuntu2204Gen2Containerd,
-			BootstrapConfigMutator: func(nbc *datamodel.NodeBootstrappingConfiguration) {
-				nbc.ContainerService.Properties.AgentPoolProfiles[0].Distro = "aks-ubuntu-containerd-22.04-gen2"
-				nbc.AgentPoolProfile.Distro = "aks-ubuntu-containerd-22.04-gen2"
-			},
-			Airgap: true,
-		},
-	}
-}
