--- e2e/scenario/scenario_ubuntu2204.go
@@ -1,27 +1,0 @@
-package scenario
-
-import (
-	"github.com/Azure/agentbaker/pkg/agent/datamodel"
-	"github.com/Azure/agentbakere2e/cluster"
-	"github.com/Azure/agentbakere2e/config"
-)
-
-func ubuntu2204() *Scenario {
-	return &Scenario{
-		Name:        "ubuntu2204",
-		Description: "Tests that a node using the Ubuntu 2204 VHD can be properly bootstrapped",
-		Tags: Tags{
-			Name:     "ubuntu2204",
-			OS:       "ubuntu2204",
-			Platform: "x64",
-		},
-		Config: Config{
-			Cluster:     cluster.ClusterKubenet,
-			VHDSelector: config.VHDUbuntu2204Gen2Containerd,
-			BootstrapConfigMutator: func(nbc *datamodel.NodeBootstrappingConfiguration) {
-				nbc.ContainerService.Properties.AgentPoolProfiles[0].Distro = "aks-ubuntu-containerd-22.04-gen2"
-				nbc.AgentPoolProfile.Distro = "aks-ubuntu-containerd-22.04-gen2"
-			},
-		},
-	}
-}
